+++ Sources/SwiftUIKit/Lists/ListBadgeIcon.swift
@@ -11,16 +11,25 @@ import SwiftUI
 
 /**
  This view mimics the color badge icons that can be found in
- e.g. System Settings.
+ e.g. System Settings lists on iOS.
 
- You can use the `.listBadgeIcon()` view modifier to convert
- any `Image` to a list badge icon.
+ You can use the `.listBadgeIcon(...)` view modifier to turn
+ any `Image` into a list badge icon.
+ 
+ Note that icon modification, like applying foreground color
+ and symbol variant, works best with SF Symbols.
  */
-@available(iOS 15.0, *)
+@available(iOS 16.0, *)
 public struct ListBadgeIcon: View {
-
+    
     /**
-     Create a system settings mimicing list icon.
+     Create a system settings-like badge icon, with a custom
+     badge color and a white, filled icon without a gradient.
+     
+     - Parameters:
+       - image: The image to use.
+       - color: The badge color to apply.
+       - height: The icon height, by default `30`.
      */
     public init(
         image: Image,
@@ -28,60 +37,122 @@ public struct ListBadgeIcon: View {
         height: CGFloat? = 30
     ) {
         self.image = image
-        self.color = color
+        self.badgeColor = color
+        self.iconColor = .white
+        self.iconGradient = false
+        self.height = height
+    }
+    
+    /**
+     Create a system settings-like badge icon, with a custom
+     badge color and a custom colored, filled icon.
+     
+     - Parameters:
+       - image: The image to use.
+       - badgeColor: The badge color to apply.
+       - iconColor: The icon color to apply.
+       - iconGradient: Whether or not to apply a gradient to the icon color. by default `false`.
+       - height: The icon height, by default `30`.
+     */
+    public init(
+        image: Image,
+        badgeColor: Color,
+        iconColor: Color,
+        iconGradient: Bool = false,
+        height: CGFloat? = 30
+    ) {
+        self.image = image
+        self.badgeColor = badgeColor
+        self.iconColor = iconColor
+        self.iconGradient = iconGradient
         self.height = height
     }
 
     private let image: Image
-    private let color: Color
+    private let badgeColor: Color
+    private let iconColor: Color
+    private let iconGradient: Bool
     private let height: CGFloat?
 
     public var body: some View {
         ZStack {
-            prefersGradient(color)
+            badgeColor
+                .asGradientBackground()
                 .aspectRatio(1, contentMode: .fit)
                 .cornerRadius(7)
             image.symbolVariant(.fill)
                 .padding(5)
                 .aspectRatio(1, contentMode: .fit)
-                .foregroundColor(color == .clear ? nil : .white)
-        }.frame(minHeight: height, maxHeight: height)
+                .foregroundColor(iconColor, gradientIf: iconGradient)
+        }
+        .backgroundStyle(badgeColor.gradient)
+        .frame(minHeight: height, maxHeight: height)
     }
 }
 
-@available(iOS 15.0, *)
-public extension Image {
-
-    /**
-     Convert the image to a system settings color badge icon.
-
-     Note that this behaves best with SF Symbols.
-     */
-    func listBadgeIcon(
-        _ color: Color,
-        height: CGFloat? = 30
-    ) -> some View {
+@available(iOS 16.0, *)
+public extension ListBadgeIcon {
+    
+    static var settings: ListBadgeIcon {
         ListBadgeIcon(
-            image: self,
-            color: color,
-            height: height
+            image: .symbol("gearshape"),
+            badgeColor: .gray,
+            iconColor: .white,
+            iconGradient: true
+        )
+    }
+    
+    static var heart: ListBadgeIcon {
+        ListBadgeIcon(
+            image: .symbol("heart"),
+            badgeColor: .hex(0xfafafa),
+            iconColor: .red,
+            iconGradient: true
+        )
+    }
+    
+    static var star: ListBadgeIcon {
+        ListBadgeIcon(
+            image: .symbol("star"),
+            badgeColor: .hex(0xfafafa),
+            iconColor: .yellow,
+            iconGradient: true
         )
     }
 }
 
-private extension View {
-
+@available(iOS 16.0, *)
+private extension Color {
+    
     @ViewBuilder
-    func prefersGradient(_ color: Color) -> some View {
-        if #available(iOS 16.0, *) {
-            Color.clear.overlay(color.gradient)
+    func asGradientBackground(
+        if condition: Bool = true
+    ) -> some View {
+        if condition {
+            Color.clear.overlay(self.gradient)
         } else {
-            color
+            self
         }
     }
 }
 
-@available(iOS 15.0, *)
+@available(iOS 16.0, *)
+private extension View {
+    
+    @ViewBuilder
+    func foregroundColor(
+        _ color: Color,
+        gradientIf condition: Bool
+    ) -> some View {
+        if condition {
+            self.foregroundStyle(color.gradient)
+        } else {
+            self.foregroundStyle(color)
+        }
+    }
+}
+
+@available(iOS 16.0, *)
 struct ListBadgeIcon_Previews: PreviewProvider {
 
     static var previews: some View {
@@ -90,19 +161,10 @@ struct ListBadgeIcon_Previews: PreviewProvider {
                 image: .symbol("exclamationmark.triangle"),
                 color: .orange
             )
-
-            Image.symbol("checkmark")
-                .listBadgeIcon(.green)
-
-            Image.symbol("checkmark")
-                .listBadgeIcon(.green)
-                .font(.body.bold())
             
-            Image.symbol("checkmark")
-                .listBadgeIcon(.hex(0xf0f0f0))
-            
-            Image.symbol("checkmark")
-                .listBadgeIcon(.hex("#f0f0f0") ?? .red)
+            ListBadgeIcon.settings
+            ListBadgeIcon.heart
+            ListBadgeIcon.star
         }
     }
 }