+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
 * `Color+Hex` has new static functions.
 * `ColorPickerBar` is a new color picker component.
 * `EditableView` is a new view protocol.
+* `ListBadgeIcon` now supports custom icon colors and icon gradients.
 * `ListHeaderIcon` is a new view.
 * `OptionalBinding` is a new func that makes it possible to pass in optional values as a non-optional binding.
 * `View` has new `prefers` extensions.